requirements.txt
asyncio>=3.4.3
aiogram>=2.25.1
httpx>=0.24.1
# Data base
gino==1.0.1
SQLAlchemy==1.3.24

python-dotenv>=0.21.1
filters>=1.3.2
asyncpg>=0.27.0